feat(decl): rename `declarative` commands namespace to `suite`
